# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReBridge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/listings": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Create Listing",
        "description": "Create the LISTING facet for a graded Item (Req 3.1, 3.5, 15.3).\n\nAn ungraded Item is rejected with 409 (grade required); an unknown Item with\n404. On success the Item transitions to LISTED and a LISTED event is emitted.",
        "operationId": "create_listing_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/listings/{item_id}": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Get Listing",
        "description": "Return the LISTING facet for an Item (Requirement 3.3).\n\nUnknown Item -> 404; an existing Item with no listing -> 404 (no listing).",
        "operationId": "get_listing_listings__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "listings"
        ],
        "summary": "Update Listing",
        "description": "Apply a partial update to an existing listing (Requirement 3.2).\n\nUnknown Item -> 404; an Item with no listing -> 404.",
        "operationId": "update_listing_listings__item_id__put",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "listings"
        ],
        "summary": "Delete Listing",
        "description": "Remove the LISTING facet for an Item (Requirement 3.4). Unknown Item -> 404.",
        "operationId": "delete_listing_listings__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/listings/{item_id}/buy": {
      "post": {
        "tags": [
          "listings"
        ],
        "summary": "Buy Listing",
        "description": "Simulated checkout for a listed Item (Requirements 15.5, 18.4).\n\nPayments are simulated in v1: no payment is processed. The purchase requires\nan existing listing (unknown Item or no listing -> 404), transitions the Item\nto SOLD, and emits a SOLD lifecycle event identifying the Item.",
        "operationId": "buy_listing_listings__item_id__buy_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/items/{item_id}/matches": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Matches",
        "description": "Return ranked buyer matches for a graded Item (G1).\n\nA 404 is raised for an unknown item (via the item-service lookup). An item\nthat exists but is not yet graded (no GRADE facet or status not GRADED)\nreturns an empty result: ``matches: []``, ``match_count_within_5km: 0``,\n``top_reason: null``. The matching engine ranks candidate buyers and the view\nis rendered with PII-free labels, distances, reasons, and intent tiers.",
        "operationId": "get_matches_items__item_id__matches_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/review/{item_id}": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Review Action",
        "description": "Confirm or override a pending grade; return the updated Item (G2).\n\nCONFIRM persists the existing grade as confirmed; OVERRIDE builds a grade\nfrom the item's current grade with the label replaced by ``override_grade``\n(carrying confidence/summary/defects/completeness). Both advance the item to\nGRADED. An action on a non-pending item -> 409; an invalid ``override_grade``\n-> 422.",
        "operationId": "review_action_review__item_id__post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemAggregateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BuyResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "SOLD"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Simulated checkout: no payment processed. SOLD event emitted."
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "order_id"
        ],
        "title": "BuyResponse",
        "description": "Simulated checkout result (Requirements 15.5, 18.4).\n\nPayments are simulated in v1: no payment is processed. The purchase\ntransitions the Item to SOLD and a SOLD lifecycle event is emitted."
      },
      "BuyerMatchModel": {
        "properties": {
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "display_label": {
            "type": "string",
            "title": "Display Label"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "match_score": {
            "type": "number",
            "title": "Match Score"
          },
          "match_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Match Reasons"
          },
          "intent_tier": {
            "type": "string",
            "title": "Intent Tier"
          }
        },
        "type": "object",
        "required": [
          "buyer_id",
          "display_label",
          "distance_km",
          "match_score",
          "intent_tier"
        ],
        "title": "BuyerMatchModel",
        "description": "A single ranked buyer match (G1).\n\n``distance_km`` and ``match_score`` are JSON numbers; ``intent_tier`` is one\nof ``HIGH``/``MEDIUM``/``LOW``; ``display_label`` is PII-free."
      },
      "CardFacetModel": {
        "properties": {
          "card_id": {
            "type": "string",
            "title": "Card Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "qr_target": {
            "type": "string",
            "title": "Qr Target"
          },
          "graded_at": {
            "type": "string",
            "title": "Graded At"
          },
          "warranty_stance": {
            "type": "string",
            "title": "Warranty Stance"
          },
          "annotated_photo_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Annotated Photo Keys"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "item_id",
          "signature",
          "qr_target",
          "graded_at",
          "warranty_stance"
        ],
        "title": "CardFacetModel",
        "description": "The CARD facet (Requirement 1.4)."
      },
...
      "CreateListingRequest": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Price"
          },
          "geohash5": {
            "type": "string",
            "title": "Geohash5"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "category",
          "price",
          "geohash5"
        ],
        "title": "CreateListingRequest",
        "description": "Create-listing body (Requirements 3.1, 3.5).\n\nListing an Item requires it to be graded; the service rejects an ungraded\nItem with a 409 grade-required error."
      },
...
      "DefectModel": {
        "properties": {
          "location": {
            "type": "string",
            "title": "Location"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "location",
          "severity"
        ],
        "title": "DefectModel"
      },
...
      "GradeFacetModel": {
        "properties": {
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "defects": {
            "items": {
              "$ref": "#/components/schemas/DefectModel"
            },
            "type": "array",
            "title": "Defects"
          },
          "completeness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompletenessModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "idem_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idem Key"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "grade",
          "confidence",
          "summary"
        ],
        "title": "GradeFacetModel",
        "description": "The GRADE facet (Requirement 1.4)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemAggregateResponse": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ItemMetaResponse"
          },
          "grade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GradeFacetModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardFacetModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionFacetModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "listing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListingFacetModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "ItemAggregateResponse",
        "description": "An Item's status plus exactly its persisted facets (Requirement 1.4)."
      },
      "ItemMetaResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "age_months": {
            "type": "integer",
            "title": "Age Months"
          },
          "context_source": {
            "type": "string",
            "title": "Context Source"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "context_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Ref"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "status",
          "category",
          "age_months",
          "context_source",
          "created_at"
        ],
        "title": "ItemMetaResponse",
        "description": "The META facet of an Item returned on create and within an aggregate."
      },
...
      "ListingResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "geohash5": {
            "type": "string",
            "title": "Geohash5"
          },
          "listed_at": {
            "type": "string",
            "title": "Listed At"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "status",
          "category",
          "price",
          "geohash5",
          "listed_at"
        ],
        "title": "ListingResponse",
        "description": "A listing returned by listing CRUD (Requirement 3.x)."
      },
...
      "MatchesResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/BuyerMatchModel"
            },
            "type": "array",
            "title": "Matches"
          },
          "match_count_within_5km": {
            "type": "integer",
            "title": "Match Count Within 5Km",
            "default": 0
          },
          "top_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Reason"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "generated_at"
        ],
        "title": "MatchesResponse",
        "description": "Ranked buyer matches for a graded Item (G1).\n\n``generated_at`` is the ISO-8601 instant the view was produced. An empty\nresult carries ``matches: []``, ``match_count_within_5km: 0``, and\n``top_reason: null``."
      },
...
      "ReviewActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "override_grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Grade"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReviewActionRequest",
        "description": "A reviewer confirm/override action (G2).\n\n``action`` is ``CONFIRM`` or ``OVERRIDE``. When ``OVERRIDE``,\n``override_grade`` must be a valid grade label; when ``CONFIRM`` it is\n``null``."
      },
...
      "UpdateListingRequest": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "geohash5": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geohash5"
          }
        },
        "type": "object",
        "title": "UpdateListingRequest",
        "description": "Partial listing update (Requirement 3.2). Unset fields are unchanged."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}